Architecture modified. Docstring for each class/object added.
