+++ app/src/main/java/com/example/financetracker/di/Error500Interceptor.kt
@@ -1,12 +1,13 @@
-package com.example.financetracker.data.di
+package com.example.financetracker.di
 
 import okhttp3.Interceptor
 import okhttp3.Protocol
 import okhttp3.Response
 import okhttp3.ResponseBody.Companion.toResponseBody
 
-
-// only for tests! to execute - add ".addInterceptor(Error500Interceptor())" inside provideOkHttpClient function
+/**
+ * Only for tests! to execute - add ".addInterceptor(Error500Interceptor())" inside provideOkHttpClient function.
+ */
 class Error500Interceptor : Interceptor {
     override fun intercept(chain: Interceptor.Chain): Response {
         return Response.Builder()
