+++ app/src/main/java/com/example/financetracker/presentation/screens/MyArticlesScreen/MyArticlesViewModel.kt
@@ -4,14 +4,21 @@ import androidx.compose.runtime.State
 import androidx.compose.runtime.mutableStateOf
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.example.financetracker.Result
-import com.example.financetracker.data.repository.ApiRepository
+import com.example.financetracker.utils.Result
+import com.example.financetracker.domain.FinanceRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.launch
 import javax.inject.Inject
 
+/**
+ * ViewModel for managing the MyArticlesScreen logic.
+ *
+ * Fetches transactions from the repository and shows UI state.
+ *
+ * @property repository The repository providing access to data.
+ */
 @HiltViewModel
-class MyArticlesViewModel @Inject constructor(private val repository: ApiRepository): ViewModel() {
+class MyArticlesViewModel @Inject constructor(private val repository: FinanceRepository): ViewModel() {
 
     private val _articlesState = mutableStateOf(MyArticlesUIState(isLoading = true))
     val articlesState: State<MyArticlesUIState> = _articlesState
