+++ app/src/main/java/com/example/financetracker/utils/NetworkMonitor.kt
@@ -1,4 +1,4 @@
-package com.example.financetracker
+package com.example.financetracker.utils
 
 import android.content.Context
 import android.net.ConnectivityManager
@@ -6,6 +6,11 @@ import android.net.NetworkCapabilities
 import dagger.hilt.android.qualifiers.ApplicationContext
 import javax.inject.Inject
 
+/**
+ * Checks the device's current network connection status.
+ *
+ * @property context Application context used to access system connectivity services.
+ */
 class NetworkMonitor @Inject constructor(@ApplicationContext private val context: Context) {
     fun hasNetworkConnection(): Boolean {
         val connectivityManager = context.getSystemService(Context.CONNECTIVITY_SERVICE) as ConnectivityManager
